+++ ColorSharpTests/tests/Illuminants/D65Test.cs
@@ -39,7 +39,7 @@ namespace Litipk.ColorSharpTests
 		[Test]
 		public void TestRGBValue()
 		{
-			var xyzWhitePoint = D65.XYZ_Sample.ToSRGB ();
+			var xyzWhitePoint = CIE_D65.XYZ_Sample.ToSRGB ();
 
 			// Because there are a lot of transformations, we can't ensure a very little delta
 			Assert.AreEqual (1.0, xyzWhitePoint.R, 0.0001);
@@ -47,7 +47,7 @@ namespace Litipk.ColorSharpTests
 			Assert.AreEqual (1.0, xyzWhitePoint.B, 0.0010);
 
 
-			var specWhitePoint = D65.spectrum_Sample.ToSRGB ();
+			var specWhitePoint = CIE_D65.spectrum_Sample.ToSRGB ();
 
 			// Because there are a lot of transformations, we can't ensure a very little delta
 			Assert.AreEqual (1.0, specWhitePoint.R, 0.0002);